:white_check_mark: added database scheduling tests
And no I did not spend 1 hour trying to fix my last test thinking it was a bad query, but no, it was just me not adding the right values to the participant list. Because adding a string to a list that only accepts uuids is definitely not working. And after figuring that out I also spend a solid 5 minutes to finally figure out that it should be participants in stead of participant. :sob: